refactor: Rename Http to ClientAPI
* Added AboutFormData.swift with Extension group
* Moved buildedFormData from Session to each extension
